1.5.0 release

+++ src/extension.ts
@@ -10,7 +10,7 @@ import { GroupListProvider } from "./treeview/GroupListProvider";
 import { TableNode } from "./treeview/TableNode";
 import { TablesListProvider } from "./treeview/TablesListProvider";
 import { DbConnectionUpdater } from "./treeview/DbConnectionUpdater";
-import { IPort, IConfig, ICommand } from "./view/app/model";
+import { IPort, IConfig } from "./view/app/model";
 import { readFile, parseOEFile } from "./common/OpenEdgeJsonReaded";
 
 import { VersionChecker } from "./view/app/Welcome/VersionChecker";
@@ -23,20 +23,21 @@ export function activate(context: vscode.ExtensionContext) {
   const versionChecker = new VersionChecker(context);
 
   if (versionChecker.isNewVersion()) {
-    // change it to .isNewVersion
     new WelcomePageProvider(context, versionChecker.versionFromPackage);
   }
 
   let allFileContent: string = "";
 
   vscode.workspace.onDidChangeConfiguration((event) => {
-    const affected = event.affectsConfiguration("ProBro.possiblePortsList");
+    const affected = event.affectsConfiguration(
+      `${Constants.globalExtensionKey}.possiblePortsList`
+    );
     if (!affected) {
       return;
     }
 
     const settingsPorts: number[] = vscode.workspace
-      .getConfiguration("ProBro")
+      .getConfiguration(Constants.globalExtensionKey)
       .get("possiblePortsList")!;
     if (settingsPorts.length === 0) {
       context.globalState.update(
@@ -136,8 +137,9 @@ export function activate(context: vscode.ExtensionContext) {
     Constants.dlc = defaultRuntime.path;
   }
 
-
-  let importConnections = vscode.workspace.getConfiguration('ProBro').get('importConnections');
+  let importConnections = vscode.workspace
+    .getConfiguration(Constants.globalExtensionKey)
+    .get("importConnections");
   let fileWatcher: vscode.FileSystemWatcher;
 
   if (importConnections) {
@@ -148,7 +150,9 @@ export function activate(context: vscode.ExtensionContext) {
     clearDatabaseConfigState();
   }
 
-  fileWatcher = vscode.workspace.createFileSystemWatcher("**/openedge-project.json")
+  fileWatcher = vscode.workspace.createFileSystemWatcher(
+    "**/openedge-project.json"
+  );
   fileWatcher.onDidChange((uri) => {
     if (importConnections) {
       createJsonDatabases(uri);
@@ -158,9 +162,14 @@ export function activate(context: vscode.ExtensionContext) {
   });
 
   vscode.workspace.onDidChangeConfiguration((event) => {
-
-    if (event.affectsConfiguration('ProBro.importConnections')) {
-      importConnections = vscode.workspace.getConfiguration('ProBro').get('importConnections');
+    if (
+      event.affectsConfiguration(
+        `${Constants.globalExtensionKey}.importConnections`
+      )
+    ) {
+      importConnections = vscode.workspace
+        .getConfiguration(Constants.globalExtensionKey)
+        .get("importConnections");
       if (importConnections) {
         vscode.workspace.findFiles("**/openedge-project.json").then((list) => {
           list.forEach((uri) => createJsonDatabases(uri));
@@ -169,11 +178,9 @@ export function activate(context: vscode.ExtensionContext) {
         clearDatabaseConfigState();
       }
     }
-
   });
 
   function createJsonDatabases(uri: vscode.Uri) {
-
     allFileContent = readFile(uri.path);
 
     const configs = parseOEFile(allFileContent);
@@ -200,7 +207,10 @@ export function activate(context: vscode.ExtensionContext) {
   }
 
   function clearDatabaseConfigState() {
-    context.workspaceState.update(`${Constants.globalExtensionKey}.dbconfig`, {});
+    context.workspaceState.update(
+      `${Constants.globalExtensionKey}.dbconfig`,
+      {}
+    );
     vscode.commands.executeCommand(
       `${Constants.globalExtensionKey}.refreshList`
     );
@@ -225,7 +235,7 @@ export function activate(context: vscode.ExtensionContext) {
   const groupListProvider = new GroupListProvider(context, tables);
   const groups = vscode.window.createTreeView(
     `${Constants.globalExtensionKey}-databases`,
-    { treeDataProvider: groupListProvider }
+    { treeDataProvider: groupListProvider, canSelectMany: true }
   );
 
   const connectionUpdater = new DbConnectionUpdater();
@@ -262,6 +272,7 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand(
       `${Constants.globalExtensionKey}.query`,
       (node: TableNode) => {
+        tablesListProvider.selectDbConfig(node);
         new QueryEditor(context, node, tablesListProvider, fieldsProvider);
       }
     )
@@ -270,8 +281,26 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.commands.registerCommand(
       `${Constants.globalExtensionKey}.deleteConnection`,
+      async (dbConnectionNode: DbConnectionNode) => {
+        const confirmation = await vscode.window.showWarningMessage(
+          `Are you sure you want to delete the connection "${dbConnectionNode.config.label}"?`,
+          { modal: true },
+          "Yes",
+          "No"
+        );
+
+        if (confirmation === "Yes") {
+          dbConnectionNode.deleteConnection(context);
+        }
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      `${Constants.globalExtensionKey}.refreshConnection`,
       (dbConnectionNode: DbConnectionNode) => {
-        dbConnectionNode.deleteConnection(context);
+        dbConnectionNode.refreshConnection(context);
       }
     )
   );
